requirements.txt
streamlit
pandas
google-cloud-storage
google-auth
python-dotenv
google-generativeai
google-genai
pydantic
